Handle single log file or directory input
The Path class from pathlib can take care of a lot of path manipulation
for us. a) Determine if the value provided is a file or directory.
b) Walk the directory and provide the list of files.

TODO: figure out if we want to add an allow-list for log extensions
(e.g. only interested in *.log or *.csv files.)